refactor(sourcing): clean sourcing module with better component categorization. Other cleanups
