+++ crates/pop-common/src/helpers.rs
@@ -40,6 +40,32 @@ pub fn get_project_name_from_path<'a>(path: &'a Path, default: &'a str) -> &'a s
 	path.file_name().and_then(|name| name.to_str()).unwrap_or(default)
 }
 
+/// Determines the target triple based on the current platform.
+pub fn target() -> Result<&'static str, Error> {
+	use std::env::consts::*;
+
+	if OS == "windows" {
+		return Err(Error::UnsupportedPlatform { arch: ARCH, os: OS });
+	}
+
+	match ARCH {
+		"aarch64" => {
+			return match OS {
+				"macos" => Ok("aarch64-apple-darwin"),
+				_ => Ok("aarch64-unknown-linux-gnu"),
+			}
+		},
+		"x86_64" | "x86" => {
+			return match OS {
+				"macos" => Ok("x86_64-apple-darwin"),
+				_ => Ok("x86_64-unknown-linux-gnu"),
+			}
+		},
+		&_ => {},
+	}
+	Err(Error::UnsupportedPlatform { arch: ARCH, os: OS })
+}
+
 #[cfg(test)]
 mod tests {
 	use super::*;
