+++ crates/pop-common/src/lib.rs
@@ -12,3 +12,14 @@ pub use git::{Git, GitHub, Release};
 pub use helpers::{get_project_name_from_path, replace_in_file};
 
 static APP_USER_AGENT: &str = concat!(env!("CARGO_PKG_NAME"), "/", env!("CARGO_PKG_VERSION"));
+
+/// Trait for observing status updates.
+pub trait Status {
+	/// Update the observer with the provided `status`.
+	fn update(&self, status: &str);
+}
+
+impl Status for () {
+	// no-op: status updates are ignored
+	fn update(&self, _: &str) {}
+}
